+++ src/controllers/import.js
@@ -15,6 +15,7 @@ import {
   ok,
 } from '@adobe/spacecat-shared-http-utils';
 import { isIsoDate, isObject, isValidUrl } from '@adobe/spacecat-shared-utils';
+import psl from 'psl';
 import { ErrorWithStatusCode } from '../support/utils.js';
 import ImportSupervisor from '../support/import-supervisor.js';
 import { ImportJobDto } from '../dto/import-job.js';
@@ -32,6 +33,20 @@ import { ImportJobDto } from '../dto/import-job.js';
  * @constructor
  */
 function ImportController(context) {
+  /**
+   * The import controller has a number of scopes that are required to access different parts of the
+   * import functionality. These scopes are used to validate the authenticated user has the required
+   * level of access.
+   * @type {{READ: 'imports.read', ALL_DOMAINS: 'imports.all_domains',
+   * READ_ALL: 'imports.read_all', WRITE: 'imports.write'}}
+   */
+  const SCOPE = {
+    READ: 'imports.read', // allows users to read the import jobs created with their API key
+    WRITE: 'imports.write', // allows users to create new import jobs
+    READ_ALL: 'imports.read_all', // allows users to view all import jobs
+    ALL_DOMAINS: 'imports.all_domains', // allows users to import across any domain
+  };
+
   const {
     dataAccess, sqs, s3, log, env, auth, attributes,
   } = context;
@@ -43,6 +58,7 @@ function ImportController(context) {
     env,
   };
 
+  const { authInfo: { profile } } = attributes;
   let importConfiguration = {};
   try {
     importConfiguration = JSON.parse(env.IMPORT_CONFIGURATION);
@@ -56,6 +72,7 @@ function ImportController(context) {
   const HEADER_ERROR = 'x-error';
   const STATUS_BAD_REQUEST = 400;
   const STATUS_ACCEPTED = 202;
+  const STATUS_UNAUTHORIZED = 401;
 
   function validateRequestData(data) {
     if (!isObject(data)) {
@@ -85,13 +102,18 @@ function ImportController(context) {
     }
   }
 
-  function validateImportApiKey(importApiKey, scopes) {
+  /**
+   * Verify that the authenticated user has the required level of access scope.
+   * @param scopes a list of scopes to validate the user has access to.
+   * @return {object} the user profile.
+   */
+  function validateAccessScopes(scopes) {
     log.debug(`validating scopes: ${scopes}`);
 
     try {
       auth.checkScopes(scopes);
     } catch (error) {
-      throw new ErrorWithStatusCode('Missing required scopes', 401);
+      throw new ErrorWithStatusCode('Missing required scopes', STATUS_UNAUTHORIZED);
     }
   }
 
@@ -108,6 +130,37 @@ function ImportController(context) {
   }
 
   /**
+   * Extract the domain from a URL.
+   * @param {string} inputUrl the URL to extract the domain from.
+   * @return {string} the domain extracted from the URL.
+   */
+  function getDomain(inputUrl) {
+    const parsedUrl = new URL(inputUrl);
+    const parsedDomain = psl.parse(parsedUrl.hostname);
+    return parsedDomain.domain; // Extracts the full domain (e.g., example.co.uk)
+  }
+
+  /**
+   * Check if the URLs in urlList belong to any of the base domains.
+   * @param {string[]} urlList the list of URLs to check.
+   * @param {string[]} baseDomainList the list of base domains to check against.
+   * @return {true} if all URLs belong to an allowed base domain
+   * @throws {ErrorWithStatusCode} if any URL does not belong to an allowed base domain
+   */
+  function isUrlInBaseDomains(urlList, baseDomainList) {
+    const invalidUrls = urlList.filter((inputUrl) => {
+      const urlDomain = getDomain(inputUrl);
+      return !baseDomainList.some((baseDomain) => urlDomain === baseDomain);
+    });
+
+    if (invalidUrls.length > 0) {
+      throw new ErrorWithStatusCode(`Invalid request: URLs not allowed: ${invalidUrls.join(', ')}`, STATUS_BAD_REQUEST);
+    }
+
+    return true;
+  }
+
+  /**
    * Create and start a new import job.
    * @param {UniversalContext} requestContext - The context of the universal serverless function.
    * @param {object} requestContext.multipartFormData - Parsed multipart/form-data request data.
@@ -120,10 +173,9 @@ function ImportController(context) {
 
     try {
       // The API scope imports.write is required to create a new import job
-      validateImportApiKey(importApiKey, ['imports.write']);
+      validateAccessScopes([SCOPE.WRITE]);
       validateRequestData(multipartFormData);
 
-      const { authInfo: { profile } } = attributes;
       let initiatedBy = {};
       if (profile) {
         initiatedBy = {
@@ -137,10 +189,33 @@ function ImportController(context) {
       const {
         urls, options, importScript, customHeaders,
       } = multipartFormData;
+
+      const scopes = profile.getScopes();
+
+      // We only check if the URLs belong to the allowed domains if the user has the write scope
+      // We do not need to check the domains for users with scope: all_domains
+      if (!scopes.some((scope) => scope.name === SCOPE.ALL_DOMAINS)) {
+        const allowedDomains = scopes
+          .filter((scope) => scope.name === SCOPE.WRITE
+              && scope.domains && scope.domains.length > 0)
+          .flatMap((scope) => scope.domains.map(getDomain));
+
+        if (allowedDomains.length === 0) {
+          throw new ErrorWithStatusCode('Missing domain information', STATUS_UNAUTHORIZED);
+        }
+
+        isUrlInBaseDomains(urls, allowedDomains);
+      }
+
+      log.info(`Creating a new import job with ${urls.length} URLs.`);
+
+      // Merge the import configuration options with the request options allowing the user options
+      // to override the defaults
       const mergedOptions = {
         ...importConfiguration.options,
         ...options,
       };
+
       const job = await importSupervisor.startNewJob(
         urls,
         importApiKey,
@@ -174,11 +249,11 @@ function ImportController(context) {
    * @returns {Promise<Response>} 200 OK with a JSON representation of the import jobs.
    */
   async function getImportJobsByDateRange(requestContext) {
-    const { startDate, endDate, importApiKey } = parseRequestContext(requestContext);
+    const { startDate, endDate } = parseRequestContext(requestContext);
     log.debug(`Fetching import jobs between startDate: ${startDate} and endDate: ${endDate}.`);
 
     try {
-      validateImportApiKey(importApiKey, ['imports.read_all']);
+      validateAccessScopes([SCOPE.READ_ALL]);
       validateIsoDates(startDate, endDate);
       const jobs = await importSupervisor.getImportJobsByDateRange(startDate, endDate);
       return ok(jobs.map((job) => ImportJobDto.toJSON(job)));
@@ -200,7 +275,7 @@ function ImportController(context) {
 
     try {
       // The API scope imports.read is required to get the import job status
-      validateImportApiKey(importApiKey, ['imports.read']);
+      validateAccessScopes([SCOPE.READ]);
       const job = await importSupervisor.getImportJob(jobId, importApiKey);
       return ok(ImportJobDto.toJSON(job));
     } catch (error) {
@@ -221,7 +296,7 @@ function ImportController(context) {
 
     try {
       // The API scope imports.read is required to get the import job status
-      validateImportApiKey(importApiKey, ['imports.read']);
+      validateAccessScopes([SCOPE.READ]);
       const job = await importSupervisor.getImportJob(jobId, importApiKey);
       const downloadUrl = await importSupervisor.getJobArchiveSignedUrl(job);
       return ok({
@@ -234,10 +309,33 @@ function ImportController(context) {
     }
   }
 
+  /**
+   * Get the progress of an import job. Results are broken down into the following:
+   * - complete: URLs that have been successfully imported.
+   * - failed: URLs that have failed to import.
+   * - pending: URLs that are still being processed.
+   * - redirected: URLs that have been redirected.
+   * @param requestContext - Context of the request.
+   * @return {Promise<Response>} 200 OK with a JSON representation of the import job progress.
+   */
+  async function getImportJobProgress(requestContext) {
+    const { jobId, importApiKey } = parseRequestContext(requestContext);
+
+    try {
+      validateAccessScopes([SCOPE.READ]);
+      const progress = await importSupervisor.getImportJobProgress(jobId, importApiKey);
+      return ok(progress);
+    } catch (error) {
+      log.error(`Failed to fetch the import job progress: ${error.message}`);
+      return createErrorResponse(error);
+    }
+  }
+
   return {
     createImportJob,
     getImportJobStatus,
     getImportJobResult,
+    getImportJobProgress,
     getImportJobsByDateRange,
   };
 }